feat: add confirm message before delete

+++ app/blogs/components/EditPublishDeleteBtn.tsx
@@ -30,6 +30,17 @@ export default function EditPublishDeleteBtn({ categoryId, blogId }: Props) {
   if (!session && !allowedNonLoginUsePanelCategory.includes(categoryId)) return null
 
   const handleDeleteOnclick = async () => {
+    const swalResult = await Swal.fire({
+      title: `Do you want to delete ${blogId}?`,
+      showDenyButton: true,
+      showCancelButton: false,
+      confirmButtonText: "Delete",
+      denyButtonText: `Cancel`
+    })
+    if (!swalResult.isConfirmed){
+      Swal.fire("Delete Canceled", "", "info")
+      return
+    }
     const res = await fetch(`/api/blogs/${blogId}`, {
       method: 'DELETE',
       headers: {
